Move firehose logic to lib to isolate store updating
