+++ src/listentogithub/app.js
@@ -4,9 +4,10 @@ var favicon = require('static-favicon');
 var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
-var github = require('octonode');
 var http = require('http');
 var debug = require('debug')('listentogithub');
+var getStore = require('./libs/firehose').getStore;
+var queryRate = require('./libs/firehose').queryRate;
 
 var app = express();
 var server = http.Server(app);
@@ -15,71 +16,14 @@ var io = require('socket.io')(server);
 
 var routes = require('./routes/index');
 
-// Set Github client id and secret, as well as github account to use for
-// oauth,in environmental variables.
-GITHUB_PAC = process.env.GITHUB_PAC;
-var client = github.client(GITHUB_PAC);
-var queryRate = (3600 * 1.0 / 5000) * 1000;
-var store = [];
-
-var getGithubEvents = function(cb) {
-    /* 
-     * Callback function should do something with the payload.
-     */
-    client.get('/events', {}, function(err, status, body, headers) {
-        var out = body.filter(function filterCb(el, i, ar) {
-                return el.type == "PushEvent";
-            }).map(function mapCb(el, i, ar) {
-                ret = { user: el.actor.login,
-                        repository: el.repo.name,
-                        id: el.id,
-                        time: el.created_at,
-                        commits: el.payload.size,
-                        commitMessages: (function(){
-                            return el.payload.commits.map(function(el, i, ar) {
-                                return { message: el.message,
-                                         url: el.url };
-
-                            });
-                        })()
-                    };
-                return ret;
-            });
-        if(cb) {
-            cb(out);
-        }
-    });
-};
-
-var mergePayloadToStore = function(store, payload) {
-    if(! store.length) {
-        return payload;
-    }
-    var ar = [];
-    var i = 0;
-    while(i < payload.length && payload[i].id != store[0].id) {
-        ar.push(payload[i]);
-        i++;
-    }
-    return ar;
-};
-
-// The interval to hit Github API.
-var interval = setInterval(function() {
-        getGithubEvents(function(payload) {
-            newData = mergePayloadToStore(store, payload);
-            store = newData.concat(store);
-            if(store.length > 50){
-                store = store.slice(0, 50);
-            }
-            io.emit("github payload", newData); 
-        });
-    }, queryRate);
+setInterval(function() {
+    io.emit("github payload", getStore());
+}, queryRate * 2);
 
 io.on('connection', function(socket) {
     console.log('a user connected');
+    io.emit("github payload", getStore());
 });
-
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
